+++ Info/Changelog.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.1.0
+
+* Moved all of the table objects into the TableObjects directory
+* Create a SQLite create table Test - Successful
+
 ## 0.0.4
 
 * Create Table Objects (Needs Testing)